refactor from parent modules into child modules
